# perf: crypto/types: use native slice index producing for loop in (*CompactBitArry).NumTrueBitsBefore